# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Requests a protected presigned URL for a product image upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400": { description: Invalid image metadata }
        "401": { description: Authentication required }
  /storage/uploads/finalize:
    post:
      tags: [Storage]
      operationId: finalizeProductImageUpload
      summary: Finalizes an uploaded product image with public read ACL and authenticated ownership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizeUploadRequest"
      responses:
        "200":
          description: Upload finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalizeUploadResponse"
        "400": { description: "Invalid path, size, MIME type, or image content" }
        "401": { description: Authentication required }
        "404": { description: Object not found }
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serves an uploaded product image
      parameters:
        - name: objectPath
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Image content
          content:
            "*/*":
              schema: { type: string, format: binary }
        "404": { description: Image not found }
        "403": { description: Object is private or access is forbidden }
...
  /market/emerging-ideas:
    get:
      operationId: getEmergingMarketIdeas
      tags: [market]
      summary: Returns ranked emerging-market ETF research ideas
      description: >
        Public, cached research ideas from a fixed emerging-market ETF universe.
        This is not investment advice. Yahoo Finance is the source and unavailable
        values are returned as null.
      parameters:
        - name: refresh
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Emerging-market ETF research ideas (at most five)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergingMarketIdeas"
        "503": { description: Emerging-market data temporarily unavailable }
  /market/search:
    get:
      operationId: searchMarketInstruments
      tags: [market]
      summary: Searches Yahoo Finance for ETFs and stocks
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 40 }
      responses:
        "200":
          description: Matching instruments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketSearchResult"
        "503": { description: Search temporarily unavailable }
        "400": { description: Invalid search query }
...
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Returns the signed-in user's watchlist
      responses:
        "200":
          description: Personal watchlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
        "401": { description: Authentication required }
    post:
      operationId: addWatchlistItem
      tags: [watchlist]
      summary: Adds a symbol to the signed-in user's watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistItemInput"
      responses:
        "201":
          description: Added item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
        "400": { description: Invalid symbol }
        "401": { description: Authentication required }
  /watchlist/radar:
    get:
      operationId: getWatchlistRadar
      tags: [watchlist]
      summary: Returns the signed-in user's calculated JB radar
      responses:
        "200":
          description: Personal JB radar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketRadar"
        "401": { description: Authentication required }
  /watchlist/{ticker}:
    delete:
      operationId: removeWatchlistItem
      tags: [watchlist]
      summary: Removes a symbol from the signed-in user's watchlist
      parameters:
        - name: ticker
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Removed }
        "400": { description: Invalid symbol }
        "401": { description: Authentication required }
...
components:
  schemas:
...
    UploadUrlRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        size: { type: number, minimum: 1, maximum: 5242880 }
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
      required: [name, size, contentType]
    UploadUrlResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]
    FinalizeUploadRequest:
      type: object
      properties:
        objectPath: { type: string, minLength: 1 }
      required: [objectPath]
    FinalizeUploadResponse:
      type: object
      properties:
        objectPath: { type: string }
      required: [objectPath]
...
    MarketRadar:
      type: object
      properties:
        marketStatus:
          type: string
        updatedAt:
          type: string
        source:
          type: string
        cached:
          type: boolean
        stale:
          type: boolean
          description: True when the response uses the last valid persisted reading because Yahoo Finance is temporarily unavailable.
        errors:
          type: array
          items: { type: string }
        assets:
          type: array
          items:
            $ref: "#/components/schemas/MarketAsset"
        summary:
          $ref: "#/components/schemas/RadarSummary"
      required: [marketStatus, updatedAt, source, cached, stale, errors, assets, summary]
    EmergingMarketIdeas:
      type: object
      description: Ranked research ideas only; never an investment recommendation.
      properties:
        updatedAt: { type: string }
        source: { type: string }
        cached: { type: boolean }
        stale: { type: boolean }
        errors:
          type: array
          items: { type: string }
        candidateUniverse:
          type: array
          items: { type: string }
        candidatesEvaluated: { type: number, minimum: 0 }
        ideas:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/EmergingMarketIdea"
      required: [updatedAt, source, cached, stale, errors, candidateUniverse, candidatesEvaluated, ideas]
    EmergingMarketIdea:
      type: object
      description: One ETF to investigate; null means Yahoo Finance did not provide enough data.
      properties:
        ticker: { type: string }
        name: { type: string }
        historySessions: { type: number, minimum: 0 }
        growthTrend:
          type: ["string", "null"]
          enum: ["Al alza", "Mixta", "A la baja", null]
        return200Sessions: { type: ["number", "null"], description: Total price return across the last 200 valid trading sessions, in percent. }
        dividendYield: { type: ["number", "null"], description: Yahoo Finance dividend yield, in percent. }
        peRatio: { type: ["number", "null"], description: Yahoo Finance trailing P/E. }
        profitMargin: { type: ["number", "null"], description: Yahoo Finance net profit margin, in percent. }
        score: { type: ["number", "null"], description: Transparent ranking score derived only from available 200-session return and trend. }
        rank: { type: ["number", "null"], minimum: 1, maximum: 5 }
        source: { type: string }
        updatedAt: { type: string }
      required: [ticker, name, historySessions, growthTrend, return200Sessions, dividendYield, peRatio, profitMargin, score, rank, source, updatedAt]
    MarketAsset:
      type: object
      properties:
        ticker: { type: string }
        name: { type: string }
        type: { type: string }
        price: { type: number }
        quoteUpdatedAt: { type: ["string", "null"] }
        change: { type: ["number", "null"] }
        rsi: { type: ["number", "null"] }
        sma20: { type: ["number", "null"] }
        sma50: { type: ["number", "null"] }
        sma100: { type: ["number", "null"] }
        sma200: { type: ["number", "null"] }
        pointsSma: { type: ["number", "null"] }
        pointsRsi: { type: ["number", "null"] }
        level: { type: ["number", "null"] }
        signal: { type: string }
        position: { type: string }
        trend: { type: string }
        historyDays: { type: number, minimum: 0, description: Trading sessions with valid daily closes returned by Yahoo Finance. }
        sparkline:
          type: array
          items: { type: number }
        history:
          description: Historical points with a valid Yahoo Finance date and close; incomplete rows are omitted and mismatched date/close arrays are rejected.
          type: array
          items:
            $ref: "#/components/schemas/MarketHistoryPoint"
      required: [ticker, name, type, price, quoteUpdatedAt, change, rsi, sma20, sma50, sma100, sma200, pointsSma, pointsRsi, level, signal, position, trend, historyDays, sparkline, history]
    MarketHistoryPoint:
      type: object
      properties:
        date: { type: string, description: Trading date in ISO 8601 calendar format }
        close: { type: number, description: Closing price for the trading date }
        sma20: { type: ["number", "null"] }
        sma50: { type: ["number", "null"] }
        sma100: { type: ["number", "null"] }
        sma200: { type: ["number", "null"] }
      required: [date, close, sma20, sma50, sma100, sma200]
    MarketSearchResult:
      type: object
      properties:
        ticker: { type: string }
        name: { type: string }
        type: { type: string }
        available: { type: boolean }
        analysisReady: { type: boolean, description: True when Yahoo Finance supplied enough daily closes for the complete JB calculation. }
        historyDays: { type: number, minimum: 0, description: Valid daily closes available from Yahoo Finance. }
      required: [ticker, name, type, available, analysisReady, historyDays]
    RadarSummary:
      type: object
      properties:
        total: { type: number }
        opportunities: { type: number }
        watch: { type: number }
        avoid: { type: number }
      required: [total, opportunities, watch, avoid]
...
    WatchlistItem:
      type: object
      properties:
        ticker: { type: string }
        createdAt: { type: string }
      required: [ticker, createdAt]
    WatchlistItemInput:
      type: object
      properties:
        ticker: { type: string, minLength: 1, maxLength: 12 }
      required: [ticker]
...