# lib/api-zod/src/generated/types/emergingMarketIdea.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmergingMarketIdeaGrowthTrend } from './emergingMarketIdeaGrowthTrend';

/**
 * One ETF to investigate; null means Yahoo Finance did not provide enough data.
 */
export interface EmergingMarketIdea {
  ticker: string;
  name: string;
  /** @minimum 0 */
  historySessions: number;
  /** @nullable */
  growthTrend: EmergingMarketIdeaGrowthTrend;
  /**
     * Total price return across the last 200 valid trading sessions
     * @nullable
     */
  return200Sessions: number | null;
  /**
     * Yahoo Finance dividend yield
     * @nullable
     */
  dividendYield: number | null;
  /**
     * Yahoo Finance trailing P/E.
     * @nullable
     */
  peRatio: number | null;
  /**
     * Yahoo Finance net profit margin
     * @nullable
     */
  profitMargin: number | null;
  /**
     * Transparent ranking score derived only from available 200-session return and trend.
     * @nullable
     */
  score: number | null;
  /**
     * @minimum 1
     * @maximum 5
     * @nullable
     */
  rank: number | null;
  source: string;
  updatedAt: string;
}