# lib/api-zod/src/generated/types/marketRadar.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAsset } from './marketAsset';
import type { RadarSummary } from './radarSummary';

export interface MarketRadar {
  marketStatus: string;
  updatedAt: string;
  source: string;
  cached: boolean;
  /** True when the response uses the last valid persisted reading because Yahoo Finance is temporarily unavailable. */
  stale: boolean;
  errors: string[];
  assets: MarketAsset[];
  summary: RadarSummary;
}